// src/navigation/RootNavigator.types.ts
export type RootStackParamList = {
  Tabs: undefined;
  Home: undefined;
  Search: undefined;
  CreateContent: undefined;
  Reels: undefined;
  Profile: undefined;
  Direct: undefined;
  Filming: undefined;
  Feed: undefined;
};

export enum NavigationScreens {
  TABS = 'Tabs',
  HOME = 'Home',
  SEARCH = 'Search',
  CREATE_CONTENT = 'CreateContent',
  REELS = 'Reels',
  PROFILE = 'Profile',
  DIRECT = 'Direct',
  FILMING = 'Filming',
  RECOMMENDATIONS_FEED = 'Feed',
}

export enum IconNames {
  HOME = 'home',
  SEARCH = 'search',
  CREATE_CONTENT = 'plusCircle',
  REELS = 'film',
  PROFILE = 'profile',
  RECOMMENDATIONS_FEED = 'Feed',
}